Add files via upload
Created for Space Eyes hacker house during 2023 Miami Hack Week, this script uses euclidean distance to determine when two objects, in a series of satellite images, are suspiciously close together. The threshold for suspiciously close is determined dynamically in each image based of the size of boundary boxes used created by PyTorch (not in script). The purpose of this script in the context of hack week was to detect and flag at sea transfers of crude oil, from taker to tanker, in order to better implement sanctions.